requirements.txt
pandas
numpy
matplotlib
seaborn
kafka-python
boto3
# Add airflow, jupyter, etc., as needed
